+++ HISTORY.md
@@ -4,6 +4,8 @@ https://github.com/josdejong/mathjs
 
 ## not yet released, version 0.6.0
 
+- Implemented chained operations via method math.chain(). For example
+  `math.chain(3).add(4).subtract(2).done()` will return `5`.
 - Implemented methods gcd and lcm.
 - Implemented method `Unit.in(unit)`, which creates a clone of the unit with a
   fixed representation. For example `math.unit('5.08 cm').in('inch')` will
@@ -17,6 +19,7 @@ https://github.com/josdejong/mathjs
   for example `math.in(math.unit('5.08 cm'), 'inch')`.
 - Split the end user documentation of the parser functions from the source files.
 - Removed function help and the built-in documentation from the core library.
+- Fixed constant i being defined as -1i instead of 1i.
 - Minor bug fixes.